Create exercise directories and docs
This commit adds a bunch of different directories, one per exercise that
will be shown during the workshop.

The 00 directory contains just scripts to get things installed. The
other directories are the actual exercises.

This is still work-in-progress, some of the READMEs are not fleshed out
yet, but it's a base to build from.

The main README was modified to include a pointer about how to install
kubectl-trace on the client side.